+++ src/elements/actions/UpdateOrderStatus.php
@@ -109,7 +109,7 @@ EOT;
         }
 
         if ($failureCount > 0) {
-            $message = Craft::t('commerce', 'Failed updating order status on {num, plural, =1{Order} other{Orders}}.', ['num' => $failureCount]);
+            $message = Craft::t('commerce', 'Failed updating order status on {num, plural, =1{order} other{orders}}.', ['num' => $failureCount]);
             if ($orderCount === $failureCount) {
                 $message = Craft::t('commerce', 'Failed to update {num, plural, =1{order status} other{order statuses}}.', ['num' => $failureCount]);
             }